See list of exports created

`kobo_exports` is a wrapper for kobotoolbox API `https://[kpi-url]/api/v2/assets/[assetid]/exports/`

Usage

kobo_exports(
  url = "eu.kobotoolbox.org",
  uname = "",
  pwd = "",
  encoding = "UTF-8",
  assetid = "",
  simplified = FALSE
)

Arguments

url

The `[url]` of kobotoolbox. Default is "eu.kobotoolbox.org".

uname

is username of your kobotoolbox account

pwd

is the password of the account

encoding

is the encoding to be used. Default is "UTF-8"

assetid

the asset id

simplified

if TRUE, then return a simple data frame with selected data, if FALSE returns the full list.

Value

The function returns a list of exports available for the account id and password entered.

Details

The function returns the export views.